requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
databricks-sqlalchemy>=1.0.0
databricks-sdk>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
tiktoken>=0.7.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
python-docx>=1.1.0
pyyaml>=6.0
python-multipart>=0.0.9
